requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url==3.1.2
Django==5.2.1
django-ckeditor==6.7.2
django-cors-headers==4.4.0
django-js-asset==3.1.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
gunicorn==25.3.0
idna==3.7
pillow==10.4.0
psycopg2-binary==2.9.11
PyJWT==2.8.0
pytz==2025.2
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.12.0